consumer-health-check-service/requirements.txt
kafka-python
flask
prometheus-flask-exporter